refactor opacity and translation UIs to work with scrolling lists.
