+++ src/Modules/opacity.scss
@@ -1,0 +1,201 @@
+:root {
+  --form-control-color: teal;
+  --form-control-disabled: #959495;
+}
+
+#lscg-layers {
+	position: absolute;
+
+	&.lscg-translate-dragging {
+		opacity: 0.2;
+		pointer-events: none;
+	}
+}
+
+#lscg-layers-container {
+	position: relative;
+	height: 100%;
+	width: 100%;
+    overflow: hidden;
+    display: flex;
+    flex-direction: column;
+	padding-top: 1rem;
+
+	.lscg-layers-body {
+		display: flex;
+		flex-direction: column;
+		gap: 0.5rem;
+		flex: 1 1 auto;
+		height: 100%;
+		color: white;
+		text-shadow: 1px 1px 0px black;
+
+		.lscg-layers-listing {
+			height: 100%;
+			overflow-y: auto;
+			display: flex;
+			flex-direction: column;
+			gap: 0.5rem;
+			padding: 0 .5rem;
+
+			.lscg-button {
+				padding: 0.5rem 0;
+				height: unset;
+				font-weight: bold;
+				cursor: pointer;
+				transition: all 0.1s linear;
+
+				&.selected {
+					background-color: teal;
+					color: white;
+					text-shadow: 1px 1px 0px black;
+				}
+
+				&:hover {
+					box-shadow: 0.2rem 0.2rem 0rem black;
+				}
+
+				&:active {
+					box-shadow: 1px 1px 0rem black;
+				}
+			}
+		}
+
+		.lscg-opacity-slider {
+			display: grid;
+
+			.lscg-opacity-slider-inputs {
+				display: flex;
+				gap: 0.5rem;
+			}
+			
+			legend {
+				font-weight: bold;
+			}
+		}
+	}
+}
+
+#lscg-layers-toolbar {
+	flex: 1 0 auto;
+	display: flex;
+	justify-content: space-between;
+}
+
+#lscg-layers-translate-toolbar {
+	display: flex;
+	justify-content: space-evenly;
+
+	.lscg-translate-direction {
+		display: flex;
+		flex-direction: column;
+    	align-items: center;
+
+		div {
+			user-select: none;
+			-webkit-user-select: none;
+			-moz-user-select: none;
+			
+			span {
+				cursor: pointer;
+				transition: text-shadow 0.1s linear;
+				&:hover {
+					text-shadow: 2px 2px 2px black;
+				}
+			}
+		}
+	}
+
+	input {
+		width: 5rem;
+	}
+
+	button {
+		background: white url("Icons/Reset.png");
+		background-size: cover;
+		width: 2rem;
+		height: 2rem;
+		border: 0.2rem black;
+		cursor: pointer;
+		transition: all 0.1s linear;
+
+		&:hover {
+			background: #00d5d5 url("Icons/Reset.png");
+			background-size: cover;
+			box-shadow: 0.2rem 0.2rem 0rem black;
+		}
+
+		&:active {
+			box-shadow: 1px 1px 0rem black;
+		}
+	}
+}
+
+.lscg-checkbox {
+	font-family: system-ui, sans-serif;
+	font-size: 1.3rem;
+	color: white;
+	text-shadow: 1px 1px 0px black;
+	font-weight: bold;
+	line-height: 1.1;
+	display: grid;
+	grid-template-columns: 1em auto;
+	gap: 0.5rem;
+	cursor: pointer;
+
+	input[type="checkbox"] {
+		/* Add if not using autoprefixer */
+		-webkit-appearance: none;
+		/* Remove most all native input styles */
+		appearance: none;
+		/* For iOS < 15 */
+		background-color: var(--form-background);
+		/* Not removed via appearance */
+		margin: 0;
+
+		font: inherit;
+		color: currentColor;
+		width: 1.1em;
+		height: 1.1em;
+		border: 0.1em solid black;
+		background-color: white;
+		transform: translateY(-0.075em);
+
+		display: grid;
+		place-content: center;
+	}
+
+	input[type="checkbox"]::before {
+		content: "";
+		width: 0.65em;
+		height: 0.65em;
+		clip-path: polygon(14% 44%, 0 65%, 50% 100%, 100% 16%, 80% 0%, 43% 62%);
+		transform: scale(0);
+		transform-origin: bottom left;
+		transition: 120ms transform ease-in-out;
+		box-shadow: inset 1em 1em var(--form-control-color);
+		/* Windows High Contrast Mode */
+		background-color: CanvasText;
+	}
+
+	input[type="checkbox"]:checked::before {
+		transform: scale(1);
+	}
+
+	input[type="checkbox"]:disabled {
+		--form-control-color: var(--form-control-disabled);
+
+		color: var(--form-control-disabled);
+		background-color: lightgrey;
+		cursor: not-allowed;
+	}
+}
+
+.lscg-checkbox + .lscg-checkbox {
+	margin-left: 0.5rem;
+}
+
+.lscg-checkbox--disabled {
+	color: var(--form-control-disabled);
+	cursor: not-allowed;
+}
